add authen to comment api and add comment api testing

+++ psu_course_review/routers/comments.py
@@ -8,6 +8,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 import math
 
 from .. import models
+from .. import deps
 
 router = APIRouter(prefix="/comments", tags=["comments"])
 
@@ -18,9 +19,20 @@ SIZE_PER_PAGE = 50
 @router.post("")
 async def create_comment(
     comment: models.CreatedComment,
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
     session: Annotated[AsyncSession, Depends(models.get_session)],
 ) -> models.Comment:
     db_comment = models.DBComment.model_validate(comment)
+
+    db_review_post = await session.get(models.DBReviewPost, comment.review_post_id)
+    if db_review_post is None:
+        raise HTTPException(status_code=404, detail="Review post not found")
+
+    db_comment.comment_author = current_user.first_name + " " + current_user.last_name
+
+    db_comment.review_post = db_review_post
+    db_comment.user = current_user
+
     session.add(db_comment)
     await session.commit()
     await session.refresh(db_comment)
@@ -64,6 +76,7 @@ async def read_comment(
     db_comment = await session.get(models.DBComment, comment_id)
     if db_comment is None:
         raise HTTPException(status_code=404, detail="Comment not found")
+
     return models.Comment.model_validate(db_comment)
 
 
@@ -71,12 +84,23 @@ async def read_comment(
 async def update_comment(
     comment_id: int,
     comment: models.UpdatedComment,
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
     session: Annotated[AsyncSession, Depends(models.get_session)],
 ) -> models.Comment:
-    data = comment.model_dump()
     db_comment = await session.get(models.DBComment, comment_id)
     if db_comment is None:
         raise HTTPException(status_code=404, detail="Comment not found")
+
+    if db_comment.user_id != current_user.id:
+        raise HTTPException(
+            status_code=403, detail="Forbidden, you are not the author of this comment"
+        )
+
+    comment.comment_author = db_comment.comment_author
+    comment.review_post_id = db_comment.review_post_id
+    comment.user_id = db_comment.user_id
+
+    data = comment.model_dump()
 
     db_comment.sqlmodel_update(data)
 
@@ -91,10 +115,16 @@ async def update_comment(
 async def delete_comment(
     comment_id: int,
     session: Annotated[AsyncSession, Depends(models.get_session)],
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
 ) -> dict:
     db_comment = await session.get(models.DBComment, comment_id)
     if db_comment is None:
         raise HTTPException(status_code=404, detail="Comment not found")
+
+    if db_comment.user_id != current_user.id:
+        raise HTTPException(
+            status_code=403, detail="Forbidden, you are not the author of this comment"
+        )
 
     await session.delete(db_comment)
     await session.commit()
